final-2/requirements.txt
Flask==3.0.0
requests
google-cloud-pubsub
flask_cors
uuid
gunicorn
datetime
bcrypt==4.2.1
beautifulsoup4==4.12.3
fastjsonschema==2.20.0
filelock==3.16.0
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-Testing==0.8.1
fqdn==1.5.1
gevent==24.10.3
geventhttpclient==2.3.1
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-pubsub==2.27.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
locust==2.32.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==3.0.2
msgpack==1.1.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.2.1
notebook_shim==0.2.4
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
overrides==7.7.0
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==5.29.1
psutil==6.0.0
psycopg2==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.10.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
requests==2.32.3
resolvelib==1.0.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.19.0
rsa==4.9
Send2Trash==1.8.3
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.36
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.3.0
tomli==2.0.1
tornado==6.4.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
uri-template==1.3.0
urllib3==2.2.3
virtualenv==20.26.4
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
widgetsnbextension==4.0.11
wrapt==1.17.0
zipp==3.21.0
zope.event==5.0
zope.interface==7.1.0